Handle errors in dynamic service resolution
