+++ api/src/main/java/graphql/nadel/hooks/ServiceExecutionHooks.java
@@ -2,6 +2,7 @@ package graphql.nadel.hooks;
 
 import graphql.GraphQLError;
 import graphql.PublicSpi;
+import graphql.execution.ExecutionStepInfo;
 import graphql.nadel.Service;
 import graphql.nadel.normalized.NormalizedQueryField;
 
@@ -53,11 +54,10 @@ public interface ServiceExecutionHooks {
      * Called to resolve the service that should be used to fetch data for a field that uses dynamic service resolution.
      *
      * @param services a list of all services registered on Nadel
-     * @param fieldName the name of the field
-     * @param arguments GraphQL arguments that were passed when querying the field
+     * @param executionStepInfo
      * @return the Service that should be used to fetch data for that field or an error that was raised when trying to resolve the service.
      */
-    default ServiceOrError resolveServiceForField(List<Service> services, String fieldName, Map<String, Object> arguments) {
+    default ServiceOrError resolveServiceForField(List<Service> services, ExecutionStepInfo executionStepInfo) {
         return null;
     }
 }
