+++ api/src/main/java/graphql/nadel/hooks/ServiceOrError.java
@@ -3,6 +3,8 @@ package graphql.nadel.hooks;
 import graphql.GraphQLError;
 import graphql.nadel.Service;
 
+import static graphql.Assert.assertFalse;
+
 /**
  * Represents either a {@link Service} or an error that was generated when trying to resolve the service.
  */
@@ -12,6 +14,9 @@ public class ServiceOrError {
 
 
     public ServiceOrError(Service service, GraphQLError error) {
+        assertFalse(service == null && error == null, () -> "Both service and error cannot be null");
+        assertFalse(service != null && error != null, () -> "Both service and error cannot be non-null");
+
         this.service = service;
         this.error = error;
     }
